Moved getResults(URL) to util.Util and minor fixes
